// app/src/engine/types/TriviaQuestion.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type TriviaQuestion = {
  text: string;
  /**
   * Two to four answers.
   */
  options: Array<string>;
  /**
   * Which answer is right.
   */
  correct: number;
  /**
   * Seconds to answer.
   */
  seconds: number;
};